+++ R/tool_transformations_collapse.R
@@ -24,6 +24,7 @@ Within.default.baseR <- plm:::Within.default
 Within.pseries.baseR <- plm:::Within.pseries
 Within.matrix.baseR  <- plm:::Within.matrix
 
+pseriesfy.baseR <- plm:::pseriesfy # in tool_pdata.frame.R
 
 ## ad 2) implement wrapper switches
 
@@ -563,6 +564,13 @@ Within.matrix.collapse <- function(x, effect, rm.null = TRUE, ...) {
 #   return(result)
 # }
 
+
+pseriesfy <- function(x,  ...) {
+  if(!isTRUE(getOption("plm.fast"))) {
+    pseriesfy.baseR(x, ...) } else {
+    if(!requireNamespace("collapse", quietly = TRUE)) stop(txt.no.collapse, call. = FALSE)
+    pseriesfy.collapse(x, ...) }
+}
 
 #' Option to Switch On/Off Fast Data Transformations
 #'